feat(services): migrate import domain (PR 9/N)
Stacked on PR #2110. Ninth and final domain — lifts the ~1,000-line
`packages/api/src/service/import.ts` orchestrator into the services
package as its own `@openstatus/services/import` domain.

### Services (`packages/services/src/import/`)

Split into focused files:

- **`schemas.ts`** — `PreviewImportInput` / `RunImportInput` zod. Provider
  discriminator + per-provider page-id fields live here; options schema
  is separately exported for callers that want to pre-validate.
- **`provider.ts`** — `createProvider` factory + `buildProviderConfig`
  reshape helper, isolated from the orchestrator so adding a provider is
  a one-file change.
- **`limits.ts`** — `addLimitWarnings` (shared by preview + run). Pure
  mutation on the `ImportSummary` argument; no writes.
- **`utils.ts`** — `clampPeriodicity` + `computePhaseStatus` helpers.
- **`phase-writers.ts`** — the seven phase writers (page / component
  groups / components / incidents / maintenances / monitors /
  subscribers). Each takes a `DB` explicitly so callers can thread a
  pre-opened tx; failing resources get `status: "failed"` with an error
  string rather than throwing.
- **`preview.ts`** — dry-run only; validates credentials, runs the
  provider with `dryRun: true`, emits warnings.
- **`run.ts`** — the orchestrator. Now owns the `pageId` ownership
  check (previously duplicated in the tRPC router) and emits exactly
  **one** `import.run` audit row regardless of outcome so partial /
  failed runs still show up in the audit signal. Deliberately *not*
  wrapped in `withTransaction` — imports can span minutes across dozens
  of writes and the existing UX is phase-level recovery.

### tRPC (`packages/api/src/router/import.ts`)

124 lines → 28 lines. The router is now a thin `previewImport` /
`runImport` wrapper; the input schemas and all validation live in the
service. The router-level `TRPCError`-throwing `pageId` ownership check
moved into `runImport` so non-tRPC callers (Slack / future) get the
same guard.

### Error shape changes

- Provider validation failure: `TRPCError("BAD_REQUEST")` →
  `ValidationError` → `TRPCError("BAD_REQUEST")`. Net-same.
- Unknown / wrong-workspace `pageId`: `TRPCError("NOT_FOUND")` →
  `NotFoundError` → `TRPCError("NOT_FOUND")`. Net-same.

### Tests

- Unit tests for `addLimitWarnings` / `clampPeriodicity` /
  `computePhaseStatus` move to `packages/services/src/import/__tests__/`.
- Router integration tests (`packages/api/src/router/import.test.ts`)
  that previously called `previewImport` / `runImport` directly to
  override workspace limits now route through `makeCaller(limitsOverride)`
  with an explicit `provider: "statuspage"` field. This also fixes four
  pre-existing TypeScript errors where those calls were missing the
  (required) provider discriminator.

### Enforcement

- Biome `noRestrictedImports` override adds `packages/api/src/router/import.ts`.
- Subpath export `@openstatus/services/import` added.
- `@openstatus/importers` added to services deps; services `tsconfig.json`
  bumped to `moduleResolution: "bundler"` so the importers package-exports
  map resolves (same setting `packages/api` already uses).

### Cleanup

Deletes `packages/api/src/service/import.ts` (1042 lines) and its test
file (463 lines). Only `telegram-updates.ts` remains in
`packages/api/src/service/` — that's slated for a follow-up PR.
