+++ packages/services/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "@openstatus/tsconfig/react-library.json",
   "include": ["src", "test"],
-  "exclude": ["dist", "build", "node_modules"]
+  "exclude": ["dist", "build", "node_modules"],
+  "compilerOptions": {
+    "moduleResolution": "bundler"
+  }
 }
